5_Pandav/backend/requirements.txt
fastapi
uvicorn[standard]
web3
langchain
langchain-google-genai
google-generativeai
python-dotenv
pydantic
pydantic-settings
httpx